Document layered file split
